# api/openapi.yaml
openapi: 3.1.0
info:
  title: Agentico Server API
  version: v1
  description: Loopback REST and SSE API for the Agentico headless runtime.
servers:
  - url: http://127.0.0.1:{port}
    variables:
      port:
        default: "0"
security:
  - bearerAuth: []
paths:
...
  /api/v1/features:
    get:
      operationId: listFeatures
      summary: List feature dashboard summaries.
      responses:
        "200":
          $ref: "#/components/responses/FeatureListResponse"
        "304":
          $ref: "#/components/responses/NotModified"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createFeature
      summary: Create a feature.
      parameters:
        - $ref: "#/components/parameters/TrustedMutationHeader"
      requestBody:
        $ref: "#/components/requestBodies/JSONMutation"
      responses:
        "201":
          $ref: "#/components/responses/ActionResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/config/runtime:
    get:
      operationId: getRuntimeConfig
      summary: Read runtime configuration.
      responses:
        "200":
          $ref: "#/components/responses/RuntimeConfigResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      operationId: patchRuntimeConfig
      summary: Patch runtime configuration.
      parameters:
        - $ref: "#/components/parameters/TrustedMutationHeader"
      requestBody:
        $ref: "#/components/requestBodies/JSONMutation"
      responses:
        "200":
          $ref: "#/components/responses/ActionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      operationId: putRuntimeConfig
      summary: Replace runtime configuration.
      parameters:
        - $ref: "#/components/parameters/TrustedMutationHeader"
      requestBody:
        $ref: "#/components/requestBodies/JSONMutation"
      responses:
        "200":
          $ref: "#/components/responses/ActionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/workspace/browse:
    get:
      operationId: browseWorkspace
      summary: Browse workspace paths.
      parameters:
        - name: path
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/WorkspaceBrowseResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/catalog/models:
    get:
      operationId: listModels
      summary: List available provider models.
      responses:
        "200":
          $ref: "#/components/responses/ModelCatalogResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/prompts:
    get:
      operationId: listPrompts
      summary: List pending prompt requests.
      responses:
        "200":
          $ref: "#/components/responses/PromptSnapshotResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/permissions:
    get:
      operationId: listPermissions
      summary: List pending permission requests.
      responses:
        "200":
          $ref: "#/components/responses/PermissionSnapshotResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/sessions:
    get:
      operationId: listSessions
      summary: List active and recent sessions.
      responses:
        "200":
          $ref: "#/components/responses/SessionListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/recovery:
    get:
      operationId: getRecoverySnapshot
      summary: Scan recoverable sessions.
      responses:
        "200":
          $ref: "#/components/responses/RecoverySnapshotResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/recovery/actions:
    post:
      operationId: executeRecoveryActions
      summary: Execute recovery actions from a recovery snapshot.
      parameters:
        - $ref: "#/components/parameters/TrustedMutationHeader"
      requestBody:
        $ref: "#/components/requestBodies/JSONMutation"
      responses:
        "200":
          $ref: "#/components/responses/ActionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/shutdown:
    post:
      operationId: shutdownRuntime
      summary: Request runtime shutdown.
      parameters:
        - $ref: "#/components/parameters/TrustedMutationHeader"
      requestBody:
        $ref: "#/components/requestBodies/JSONMutation"
      responses:
        "200":
          $ref: "#/components/responses/ActionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/events:
    get:
      operationId: streamEvents
      summary: Stream ordered runtime events as SSE.
      description: Emits envelope JSON under SSE `data:` lines. Every non-control event includes a monotonic `seq`, `epoch`, resource identity, and resource version. Clients resume with `Last-Event-ID` or `after`.
      security:
        - bearerAuth: []
        - sseAccessToken: []
      parameters:
        - name: after
          in: query
          required: false
          schema:
            type: integer
            format: int64
            x-go-type: uint64
            minimum: 0
        - name: epoch
          in: query
          required: false
          schema:
            type: string
        - name: heartbeat_ms
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          $ref: "#/components/responses/EventStream"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: instance-token
      description: Per-instance token from discovery metadata.
    sseAccessToken:
      type: apiKey
      in: query
      name: access_token
      description: SSE fallback for browser EventSource clients that cannot set Authorization.
  parameters:
...
    TrustedMutationHeader:
      name: X-Agentico-Client
      in: header
      required: true
      schema:
        type: string
        enum: ["local"]
      description: CSRF defense-in-depth for local browser-origin mutations. Bearer auth is still required.
  requestBodies:
    JSONMutation:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true
  responses:
...
    FeatureListResponse:
      description: Feature list.
      headers:
        X-Agentico-Seq:
          $ref: "#/components/headers/Sequence"
        ETag:
          $ref: "#/components/headers/ETag"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/FeatureListResponse"
...
    RuntimeConfigResponse:
      description: Runtime config.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RuntimeConfigResponse"
    WorkspaceBrowseResponse:
      description: Workspace browse result.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/WorkspaceBrowseResponse"
    ModelCatalogResponse:
      description: Model catalog.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ModelCatalogResponse"
    PromptSnapshotResponse:
      description: Pending prompt snapshot.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PromptSnapshotResponse"
    PermissionSnapshotResponse:
      description: Pending permission snapshot.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PermissionSnapshotResponse"
...
    SessionListResponse:
      description: Session list.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SessionListResponse"
...
    RecoverySnapshotResponse:
      description: Recovery snapshot.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RecoverySnapshotResponse"
    ActionResponse:
      description: Mutation result.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ActionResponse"
    ErrorResponse:
      description: Structured API error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotModified:
      description: Snapshot revision matches the supplied validator.
    EventStream:
      description: Runtime SSE stream carrying `SSEEvent` JSON envelopes.
      content:
        text/event-stream:
          schema:
            $ref: "#/components/schemas/SSEEvent"
...
  headers:
    Sequence:
      description: Current event broker sequence at response time.
      schema:
        type: integer
        minimum: 0
    ETag:
      description: Snapshot revision validator.
      schema:
        type: string
  schemas:
...
    ErrorResponse:
      type: object
      required: [api_version, error]
      properties:
        api_version:
          type: string
        error:
          $ref: "#/components/schemas/Error"
...
    FeatureListResponse:
      allOf:
        - $ref: "#/components/schemas/JSONResponse"
        - type: object
          required: [features]
          properties:
            features:
              type: array
              items:
                $ref: "#/components/schemas/FeatureSummary"
            warnings:
              type: array
              items:
                $ref: "#/components/schemas/Warning"
...
    RuntimeConfigResponse:
      allOf:
        - $ref: "#/components/schemas/JSONResponse"
        - type: object
          required:
            - runtime
            - model_defaults
            - feature_defaults
            - repos
            - ui
            - notifications
            - observability
            - providers
          properties:
            runtime:
              $ref: "#/components/schemas/RuntimeIdentity"
            model_defaults:
              $ref: "#/components/schemas/ModelDefaults"
              x-go-name: Defaults
            feature_defaults:
              $ref: "#/components/schemas/FeatureDefaults"
            repos:
              type: array
              items:
                $ref: "#/components/schemas/ConfigRepo"
            workspace_roots:
              type: array
              items:
                type: string
            ui:
              x-go-type: config.UIConfig
              x-go-type-import:
                path: github.com/doordash-oss/agentic-orchestrator/internal/config
                name: config
            notifications:
              $ref: "#/components/schemas/NotificationConfig"
            observability:
              $ref: "#/components/schemas/Observability"
            providers:
              type: array
              items:
                type: string
    WorkspaceBrowseResponse:
      allOf:
        - $ref: "#/components/schemas/JSONResponse"
        - type: object
          required: [path, is_git_repo, child_repo_count, entries]
          properties:
            path:
              type: string
            is_git_repo:
              type: boolean
            child_repo_count:
              type: integer
            entries:
              type: array
              items:
                $ref: "#/components/schemas/WorkspaceBrowseEntry"
    ModelCatalogResponse:
      allOf:
        - $ref: "#/components/schemas/JSONResponse"
        - type: object
          required: [provider_order, provider_models, phase_defaults, phase_provider_models]
          properties:
            provider_order:
              type: array
              items:
                type: string
            provider_models:
              type: object
              additionalProperties:
                type: array
                items:
                  $ref: "#/components/schemas/Model"
            phase_defaults:
              $ref: "#/components/schemas/ModelDefaults"
            phase_provider_models:
              type: object
              additionalProperties:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
    PromptSnapshotResponse:
      allOf:
        - $ref: "#/components/schemas/JSONResponse"
        - type: object
          required: [ask_user_questions, help_queue, need_user_inputs]
          properties:
            ask_user_questions:
              type: array
              items:
                $ref: "#/components/schemas/ControlRequest"
            help_queue:
              type: array
              items:
                $ref: "#/components/schemas/HelpQueue"
            need_user_inputs:
              type: array
              items:
                $ref: "#/components/schemas/NeedUserInputGate"
    PermissionSnapshotResponse:
      allOf:
        - $ref: "#/components/schemas/JSONResponse"
        - type: object
          required: [requests]
          properties:
            requests:
              type: array
              items:
                $ref: "#/components/schemas/ControlRequest"
...
    SessionListResponse:
      allOf:
        - $ref: "#/components/schemas/JSONResponse"
        - type: object
          required: [sessions]
          properties:
            sessions:
              type: array
              items:
                $ref: "#/components/schemas/SessionSummary"
...
    ActionResponse:
      allOf:
        - $ref: "#/components/schemas/ActionBaseResponse"
        - type: object
          properties:
            create_feature_response:
              $ref: "#/components/schemas/CreateFeatureResponse"
            feature_start_response:
              $ref: "#/components/schemas/FeatureStartResponse"
            feature_stop_response:
              $ref: "#/components/schemas/FeatureStopResponse"
            feature_restart_response:
              $ref: "#/components/schemas/FeatureRestartResponse"
            review_decision_response:
              $ref: "#/components/schemas/ReviewDecisionResponse"
            feature_config_update_response:
              $ref: "#/components/schemas/FeatureConfigUpdateResponse"
            need_user_input_decision_response:
              $ref: "#/components/schemas/NeedUserInputDecisionResponse"
            need_user_input_draft_response:
              $ref: "#/components/schemas/NeedUserInputDraftResponse"
            input_notifications_toggle_response:
              $ref: "#/components/schemas/InputNotificationsToggleResponse"
            permission_answer_response:
              $ref: "#/components/schemas/PermissionAnswerResponse"
            ask_user_answer_response:
              $ref: "#/components/schemas/AskUserAnswerResponse"
            help_send_response:
              $ref: "#/components/schemas/HelpSendResponse"
            chat_start_response:
              $ref: "#/components/schemas/ChatStartResponse"
            runtime_config_update_response:
              $ref: "#/components/schemas/RuntimeConfigUpdateResponse"
            publish_feature_response:
              $ref: "#/components/schemas/PublishFeatureResponse"
            publish_description_response:
              $ref: "#/components/schemas/PublishDescriptionResponse"
            merge_feature_response:
              $ref: "#/components/schemas/MergeFeatureResponse"
            rewind_feature_response:
              $ref: "#/components/schemas/RewindFeatureResponse"
            retry_feature_response:
              $ref: "#/components/schemas/RetryFeatureResponse"
            rebase_start_response:
              $ref: "#/components/schemas/RebaseStartResponse"
            review_comments_fetch_response:
              $ref: "#/components/schemas/ReviewCommentsFetchResponse"
            review_comments_start_response:
              $ref: "#/components/schemas/ReviewCommentsStartResponse"
            tweak_start_response:
              $ref: "#/components/schemas/TweakStartResponse"
            tweak_finish_response:
              $ref: "#/components/schemas/TweakFinishResponse"
            refactor_start_response:
              $ref: "#/components/schemas/RefactorStartResponse"
            refactor_restart_response:
              $ref: "#/components/schemas/RefactorRestartResponse"
            mark_done_response:
              $ref: "#/components/schemas/MarkDoneResponse"
            cleanup_feature_response:
              $ref: "#/components/schemas/CleanupFeatureResponse"
            delete_feature_response:
              $ref: "#/components/schemas/DeleteFeatureResponse"
            recovery_action_response:
              $ref: "#/components/schemas/RecoveryActionResponse"
            shutdown_response:
              $ref: "#/components/schemas/ShutdownResponse"
...